-Implementando mais métodos

+++ src/main/java/br/com/dextra/bootcamp/MentorMatch/repository/MentorRepository.java
@@ -3,5 +3,5 @@ package br.com.dextra.bootcamp.MentorMatch.repository;
 import br.com.dextra.bootcamp.MentorMatch.model.Mentor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MentorRepository extends JpaRepository<Mentor, Integer> {
+public interface MentorRepository extends JpaRepository<Mentor, Long> {
 }
